float offset

+++ lib/android/src/main/java/com/airbnb/android/react/maps/AirMapManager.java
@@ -236,7 +236,7 @@ public class AirMapManager extends ViewGroupManager<AirMapView> {
     float bearing;
     float angle;
     float altitude;
-    Double offsetMeters;
+    float offsetMeters;
     ReadableMap region;
 
     switch (commandId) {
